Waste_detection/requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
fastapi
uvicorn
python-multipart
pillow